+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+
+## 1.1.0
+
+- Added plural block for `two`.
+- Added description to messages, using `##` delimiter at end of message value.
+
 ## 1.0.0
 
 - Initial version, created by Stagehand